// src/types.ts
export interface GalleryItem {
  id: string;
  url: string;
  alt: string;
  category: string;
  title: string;
}

export interface AmenityItem {
  id: string;
  title: string;
  description: string;
  iconName: string;
}

export interface WhyChooseItem {
  id: string;
  title: string;
  description: string;
}